// types/types.d.ts
export interface NacbarCardData {
    title: string;
    icon?: string;
    description: string;
}

export interface NavbarBlogData {
    date: string;
    title: string;
    description: string;
    link: string;
}

export interface NavbarcallsToActionData {
    name: string;
    href: string;
    icon: React.ElementType;
}

export interface NavbarProductData {
    name: string;
    description: string;
    href: string;
    icon: React.ElementType;
}